- added collection_slug and banner_image_url to OpenSeaCollectionMetadata - added collection and mint to Nft - updated NftSaleFeeData - updated NftSale - changed ValidAt structure - fixed get_contract_metadata_batch test

+++ alchemy/nft/raw.py
@@ -125,7 +125,7 @@ class RawReingestContractResponse(TypedDict):
 class RawNftAttributeRarity(TypedDict):
     value: str
     trait_type: str
-    prevalence: int
+    prevalence: float
 
 
 class RawComputeRarityResponse(TypedDict):
